agrishield-backend/requirements/base.txt
# Django & Core
Django>=4.2,<5.0

# REST API
djangorestframework>=3.14
django-cors-headers>=4.0

# Env vars
python-dotenv>=1.0.1

# Database
psycopg2-binary>=2.9  # PostgreSQL driver (optional, remove if SQLite only)

# Africa's Talking SDK (SMS, USSD)
africastalking>=1.2.5

# Utilities
requests>=2.31
